05_django/04_drf_auth/02_homework/django_hw_4_4/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2025.4.26
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==5.1.0
Django==3.2.25
django-allauth==0.53.1
django-cors-headers==3.10.1
djangorestframework==3.15.1
idna==3.10
importlib-resources==5.4.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2025.2
requests==2.27.1
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.20
zipp==3.6.0
